GPX writing is in seperate extension
Fitdataprotocol pod is added
Some headers are added
