+++ GpxExport/Models/Workout.swift
@@ -13,9 +13,9 @@ import MapKit
 
 struct Workout {
     private var hkWorkout: HKWorkout
-    private var route: [CLLocation]
-    private var heartRate: [HKQuantitySample]
-    private var startDate: Date
+    var route: [CLLocation]
+    var heartRate: [HKQuantitySample]
+    var startDate: Date
 
     var activityType: String {
         return hkWorkout.formattedWorkoutType
@@ -82,61 +82,10 @@ struct Workout {
     }
 
     func writeFile() -> URL? {
-        var currentHeartrateIndex = 0
-        var currentHeartrate: Double = -1
-        let bpmUnit = HKUnit(from: "count/min")
-        var heartrateString = ""
-
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH.mm.ss"
-
-        let fileName = "\(formatter.string(from: startDate)) - \(activityType)"
-
-        let targetURL = URL(fileURLWithPath: NSTemporaryDirectory())
-            .appendingPathComponent(fileName)
-            .appendingPathExtension("gpx")
-
-        DispatchQueue.global(qos: .background).async {
-            let file: FileHandle
-
-            do {
-                let manager = FileManager.default
-                if manager.fileExists(atPath: targetURL.path) {
-                    try manager.removeItem(atPath: targetURL.path)
-                }
-                manager.createFile(atPath: targetURL.path, contents: Data())
-                file = try FileHandle(forWritingTo: targetURL)
-            } catch let err {
-                print(err)
-                return
-            }
-
-            if let header = self.gpxHeader(title: self.name, startDate: self.startDate).data(using: .utf8) {
-                file.write(header)
-            }
-
-            for location in self.route {
-                while (currentHeartrateIndex < self.self.heartRate.count) && (location.timestamp > self.heartRate[currentHeartrateIndex].startDate) {
-                    currentHeartrate = self.heartRate[currentHeartrateIndex].quantity.doubleValue(for: bpmUnit)
-                    currentHeartrateIndex += 1
-                    heartrateString = self.gpxHeartRate(currentHeartrate)
-                }
-                if let trackpoint = self.gpxTrackPoint(location: location, heartrate: heartrateString).data(using: .utf8) {
-                    file.write(trackpoint)
-                }
-            }
-            file.write("""
-  </trkseg>
-  </trk>
-  </gpx>
-
-  """.data(using: .utf8)!)
-            file.closeFile()
-        }
-        return targetURL
+        return writeGPX()
     }
 
-    private func gpxTrackPoint(location: CLLocation, heartrate: String) -> String {
+    func gpxTrackPoint(location: CLLocation, heartrate: String) -> String {
         let isoFormatter = ISO8601DateFormatter()
 
         return """
@@ -148,7 +97,7 @@ struct Workout {
         """
     }
 
-    private func gpxHeartRate(_ currentHeartrate: Double) -> String {
+    func gpxHeartRate(_ currentHeartrate: Double) -> String {
         return """
 
         <extensions>
@@ -158,7 +107,7 @@ struct Workout {
         </extensions>
         """
     }
-    private func gpxHeader(title: String, startDate: Date) -> String {
+    func gpxHeader(title: String, startDate: Date) -> String {
         let isoFormatter = ISO8601DateFormatter()
 
         let formatter = DateFormatter()
