+++ GpxExport/Models/WorkoutDataStore.swift
@@ -1,3 +1,11 @@
+//
+//  WorkoutDataStore.swift
+//  GpxExport
+//
+//  Created by Mario Martelli on 30.11.17.
+//  Copyright © 2017 Mario Martelli. All rights reserved.
+//
+
 import HealthKit
 import WatchKit
