+++ GpxExport/Utilities/HealthKitSetupAssistant.swift
@@ -1,3 +1,11 @@
+//
+//  Workout.swift
+//  GpxExport
+//
+//  Created by Mario Martelli on 30.11.17.
+//  Copyright © 2017 Mario Martelli. All rights reserved.
+//
+
 import HealthKit
 
 class HealthKitSetupAssistant {
