+++ Podfile
@@ -1,4 +1,5 @@
 platform :ios, '11.0'
 target 'GpxExport' do
   pod 'SwiftLint'
+  pod 'FitDataProtocol'
 end
